Refactor PlayerScore into Score
Remove the SCSS related to the scores of previous players as it was duplicating with the SCSS of PlayerScore.

Refactor the component by removing Redux from within it and using a separate prop called 'score' independent of the state management.

<PlayerScore/> becomes <Score/> and handles both the scores of previous players and the current player based on the 'isItActualPlayer' prop (true if it's the current player, false if it's not the current player).
