+++ src/components/Score/Score.scss
@@ -1,18 +1,19 @@
-.player_score {
+.scores {
 	display: flex;
 	min-width: 200px;
 	justify-content: space-between;
 
-	> .score,
-	> .name {
+	.score,
+	.name {
 		display: flex;
 		justify-content: space-between;
 	}
 
-	> .name {
+	.name {
 		color: white;
 		border: none;
 		outline: none;
+		overflow: hidden;
 		background-color: black;
 		max-width: 150px;
 		font-size: 30px;
@@ -32,5 +33,4 @@
 			}
 		}
 	}
-
 }
