+++ src/scene/HighScore.scss
@@ -20,35 +20,7 @@
 		flex-direction: row;
 	}
 
-	.scores {
-		display: flex;
-		min-width: 200px;
-		justify-content: space-between;
-
-		>.score,
-		>.name {
-			display: flex;
-			justify-content: space-between;
-			animation: opacity 2.5s forwards;
-
-			@keyframes opacity {
-				from {
-					opacity: 0;
-				}
-
-				to {
-					opacity: 1;
-				}
-			}
-		}
-
-		>.name {
-			max-width: 150px;
-			white-space: nowrap;
-			overflow: hidden;
-		}
-
-	}
+	
 
 	@mixin changeColorInfinite($fromColor, $toColor) {
 		animation: color-#{$fromColor}-#{$toColor} 0.75s infinite alternate;
